# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  user.CreateUserDto:
    properties:
      email:
        type: string
      password:
        minLength: 8
        type: string
      permissions:
        items:
          type: string
        type: array
      username:
        maxLength: 30
        minLength: 3
        type: string
    required:
    - email
    - password
    type: object
  user.DeleteUsersDto:
    properties:
      ids:
        items:
          type: string
        type: array
    type: object
  user.MeResponseDto:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      permissions:
        items:
          type: string
        type: array
      username:
        type: string
    type: object
  user.PermissionGroupResponseDto:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
      permissions:
        items:
          $ref: '#/definitions/user.PermissionResponseDto'
        type: array
    type: object
  user.PermissionResponseDto:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  user.UpdateUserDto:
    properties:
      email:
        type: string
      permissions:
        items:
          type: string
        type: array
      username:
        maxLength: 30
        minLength: 3
        type: string
    type: object
  user.UserResponseDto:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      permission_groups:
        items:
          $ref: '#/definitions/user.PermissionGroupResponseDto'
        type: array
      username:
        type: string
    type: object
  utils.ErrorResponse:
    properties:
      error:
        example: Error details
        type: string
      message:
        example: Operation failed
        type: string
      success:
        example: false
        type: boolean
    type: object
  utils.SuccessResponse:
    properties:
      data: {}
      length:
        type: integer
      message:
        example: Operation successful
        type: string
      success:
        example: true
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@noheir.com
    name: API Support
  description: API server for NoHeir application
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: NoHeir API
  version: "1.0"
paths:
...
  /users:
    delete:
      consumes:
      - application/json
      description: Delete multiple users by their UUIDs
      parameters:
      - description: Array of user IDs to delete
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.DeleteUsersDto'
      produces:
      - application/json
      responses:
        "200":
          description: Users deleted successfully
          schema:
            $ref: '#/definitions/utils.SuccessResponse'
        "400":
          description: Invalid request body or invalid user ID
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "403":
          description: Authorization failed
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Failed to delete users
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - SessionAuth: []
      summary: Delete users
      tags:
      - Users
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of users with optional search filter
      parameters:
      - description: Search term to filter users by username or email
        in: query
        name: search
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Users fetched successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/user.UserResponseDto'
                  type: array
              type: object
        "400":
          description: Invalid pagination parameters
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "403":
          description: Authorization failed
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Failed to get all users
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - SessionAuth: []
      summary: Get all users
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Create a new user with username, email, password, and optional
        permissions
      parameters:
      - description: User creation details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.CreateUserDto'
      produces:
      - application/json
      responses:
        "201":
          description: User created successfully
          schema:
            $ref: '#/definitions/utils.SuccessResponse'
        "400":
          description: Invalid username or invalid permissions
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "403":
          description: Authorization failed
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "409":
          description: Username or email already exists
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "422":
          description: Validation failed
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Failed to create user
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - SessionAuth: []
      summary: Create a new user
      tags:
      - Users
  /users/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve a single user by their UUID
      parameters:
      - description: User ID (UUID format)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User fetched successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/user.UserResponseDto'
              type: object
        "400":
          description: Invalid ID format
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "403":
          description: Authorization failed
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Failed to get user
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - SessionAuth: []
      summary: Get user by ID
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Update an existing user by their UUID
      parameters:
      - description: User ID (UUID format)
        in: path
        name: id
        required: true
        type: string
      - description: Updated user details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.UpdateUserDto'
      produces:
      - application/json
      responses:
        "200":
          description: User updated successfully
          schema:
            $ref: '#/definitions/utils.SuccessResponse'
        "400":
          description: Invalid user ID, invalid username, or invalid permissions
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "403":
          description: Authorization failed
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "422":
          description: Validation failed
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Failed to update user
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - SessionAuth: []
      summary: Update a user
      tags:
      - Users
  /users/me:
    get:
      consumes:
      - application/json
      description: Retrieve the currently authenticated user's information with their
        permissions
      produces:
      - application/json
      responses:
        "200":
          description: User fetched successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/user.MeResponseDto'
              type: object
        "400":
          description: Invalid ID format
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Failed to get me
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - SessionAuth: []
      summary: Get current user
      tags:
      - Users
schemes:
- http
- https
securityDefinitions:
  SessionAuth:
    in: cookie
    name: auth_session
    type: apiKey
swagger: "2.0"